Rename file from Narrow-Sidebar.js to NarrowSidebar.js and update imports
