+++ labsystem/src/Sidebar/NarrowSidebar.js
