+++ stories/NarrowSidebar.stories.mdx
@@ -1,7 +1,7 @@
 import { Fragment } from 'react';
 import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
 
-import NarrowSidebar from '../labsystem/src/Sidebar/Narrow-Sidebar';
+import NarrowSidebar from '../labsystem/src/Sidebar/NarrowSidebar';
 import Header from '../labsystem/src/Sidebar/Header';
 import Body from '../labsystem/src/Sidebar/Body';
 import Footer from '../labsystem/src/Sidebar/Footer';
